backend/requirements.txt
# FastAPI and server
fastapi>=0.115.0
uvicorn[standard]>=0.34.0

# Pydantic for data validation (>=2.11.7 required by supabase/realtime)
pydantic>=2.11.7
pydantic-settings>=2.7.0

# Supabase
supabase>=2.27.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# HTTP client (supabase 2.27 requires httpx>=0.26,<0.29)
httpx>=0.28.0,<0.29

# Environment variables
python-dotenv>=1.0.1

# JSON schema validation
jsonschema>=4.23.0

# Utilities
python-slugify>=8.0.1

# Development
pytest>=8.3.0
pytest-asyncio>=0.25.0
